송민규/마일스톤조회성공/260623

+++ src/main/java/com/weple/cloud/milestone/service/MilestoneVO.java
@@ -21,8 +21,8 @@ import lombok.ToString;
 @Builder
 @AllArgsConstructor
 public class MilestoneVO {
-	private Integer milestoneId;
-	private Integer projectId;
+	private Long milestoneId;
+	private Long projectId;
 	private String  userCode;
 	private String  milestoneTitle;
 	private String  milestoneDescribe;
